refactor(ui): decompose ProductivityScreen into modular components
Extracted monolithic 858-line `productivity.tsx` into discrete, reusable UI components (`ProductivityHeader`, `ProductivityTabs`, `TaskPriorityFilters`, `TodayPlanCard`, `ProductivityEmptyState`). Moved view-model state orchestration, filtering logic, and dependency hook integrations out of the UI tree and into a dedicated `useProductivityViewModel` custom hook to strictly adhere to SOLID and Clean Architecture principles. Fixes all related eslint warnings for unused imports.
